Implement Golomb-Rice filtering backend call

+++ MagicalCryptoWallet.Tests/WalletTests.cs
@@ -1,13 +1,17 @@
-﻿using MagicalCryptoWallet.KeyManagement;
+﻿using MagicalCryptoWallet.Backend.Models;
+using MagicalCryptoWallet.KeyManagement;
 using MagicalCryptoWallet.Logging;
 using MagicalCryptoWallet.Models;
 using MagicalCryptoWallet.Services;
+using MagicalCryptoWallet.Tests.NodeBuilding;
 using NBitcoin;
 using NBitcoin.Protocol;
 using NBitcoin.Protocol.Behaviors;
+using NBitcoin.RPC;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -52,7 +56,8 @@ namespace MagicalCryptoWallet.Tests
 			var dataFolder = Path.Combine(SharedFixture.DataDir, nameof(TestServicesAsync));
 			Directory.CreateDirectory(SharedFixture.DataDir);
 
-			var addressManagerFilePath = Path.Combine(SharedFixture.DataDir, $"AddressManager{network}.dat");
+			var addressManagerFolderPath = Path.Combine(SharedFixture.DataDir, "AddressManager");
+			var addressManagerFilePath = Path.Combine(addressManagerFolderPath, $"AddressManager{network}.dat");
 			var blocksFolderPath = Path.Combine(SharedFixture.DataDir, $"Blocks{network}");
 			var connectionParameters = new NodeConnectionParameters();
 			AddressManager addressManager = null;
@@ -63,6 +68,12 @@ namespace MagicalCryptoWallet.Tests
 				{
 					addressManager = AddressManager.LoadPeerFile(addressManagerFilePath);
 					Logger.LogInfo<WalletService>($"Loaded {nameof(AddressManager)} from `{addressManagerFilePath}`.");
+				}
+				catch(DirectoryNotFoundException ex)
+				{
+					Logger.LogInfo<WalletService>($"{nameof(AddressManager)} did not exist at `{addressManagerFilePath}`. Initializing new one.");
+					Logger.LogTrace<WalletService>(ex);
+					addressManager = new AddressManager();
 				}
 				catch (FileNotFoundException ex)
 				{
@@ -152,10 +163,14 @@ namespace MagicalCryptoWallet.Tests
 				// So next test will download the block.
 				foreach (var hash in blocksToDownload) 
 				{
-					downloader.TryRemove(hash);
+					downloader?.TryRemove(hash);
 				}
-				Directory.Delete(blocksFolderPath, recursive: true);
+				if (Directory.Exists(blocksFolderPath))
+				{
+					Directory.Delete(blocksFolderPath, recursive: true);
+				}
 
+				Directory.CreateDirectory(Path.GetDirectoryName(addressManagerFilePath));
 				addressManager?.SavePeerFile(addressManagerFilePath, network);
 				Logger.LogInfo<WalletTests>($"Saved {nameof(AddressManager)} to `{addressManagerFilePath}`.");
 			}
@@ -181,6 +196,75 @@ namespace MagicalCryptoWallet.Tests
 		{
 			Interlocked.Increment(ref _mempoolTransactionCount);
 			Logger.LogInfo<WalletTests>($"Mempool transaction received: {e.GetHash()}.");
+		}
+
+		[Fact]
+		public async Task FilterBuilderTestAsync()
+		{
+			using (var builder = NodeBuilder.Create())
+			{
+				builder.CreateNode();
+				builder.StartAll();
+				CoreNode regtestNode = builder.Nodes[0];
+				regtestNode.Generate(101);
+				RPCClient rpc = regtestNode.CreateRPCClient();
+
+				var indexBuilderServiceDir = Path.Combine(SharedFixture.DataDir, nameof(IndexBuilderService));
+				var indexFilePath = Path.Combine(indexBuilderServiceDir, $"Index{rpc.Network}.dat");
+				var utxoSetFilePath = Path.Combine(indexBuilderServiceDir, $"UtxoSet{rpc.Network}.dat");
+
+				var indexBuilderService = new IndexBuilderService(rpc, indexFilePath, utxoSetFilePath);
+				try
+				{
+					indexBuilderService.Syncronize();
+
+					// Test initial syncronization.
+					var times = 0;
+					uint256 firstHash = await rpc.GetBlockHashAsync(0);
+					while (indexBuilderService.GetFilters(firstHash).Count() != 101)
+					{
+						if (times > 500) // 30 sec
+						{
+							throw new TimeoutException($"{nameof(IndexBuilderService)} test timed out.");
+						}
+						await Task.Delay(100);
+						times++;
+					}
+
+					// Test later syncronization.
+					regtestNode.Generate(10);
+					times = 0;
+					while (indexBuilderService.GetFilters(firstHash).Count() != 111)
+					{
+						if (times > 500) // 30 sec
+						{
+							throw new TimeoutException($"{nameof(IndexBuilderService)} test timed out.");
+						}
+						await Task.Delay(100);
+						times++;
+					}
+
+					// Test correct number of filters is received.
+					var hundredthHash = await rpc.GetBlockHashAsync(100);
+					Assert.Equal(11, indexBuilderService.GetFilters(hundredthHash).Count());
+					var bestHash = await rpc.GetBestBlockHashAsync();
+					Assert.Empty(indexBuilderService.GetFilters(bestHash));
+
+					// Test filter block hashes are correct.
+					var filters = indexBuilderService.GetFilters(firstHash).ToArray();
+					for (int i = 0; i < 111; i++)
+					{
+						var expectedHash = await rpc.GetBlockHashAsync(i + 1);
+						var filterModel = FilterModel.FromLine(filters[i], new Height(i));
+						Assert.Equal(expectedHash, filterModel.BlockHash);
+						Assert.Null(filterModel.Filter);
+					}
+				}
+				finally
+				{
+					indexBuilderService.Stop();
+				}
+			}
 		}
 	}
 }
